requirements.txt
Flask
opencv-python
numpy
gunicorn
Pillow
torch
